requirements.txt
torch==2.8.0
torchvision==0.16.1
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.0.0
opencv-python==4.8.0.74
numpy==1.24.3
Werkzeug==3.0.0
